Fix(same comment on same highlight):Same user should not comment on same highlight twice  - A user cannot post same comment on a hightlight twice  - Breakdown views into different files [Delivers #163634691]
