+++ authors/apps/articles/utils.py
@@ -1,0 +1,15 @@
+from rest_framework.exceptions import ValidationError
+from.models import Article
+
+
+def article_not_found():
+    raise ValidationError(
+        detail={'error': 'No article found for the slug given'})
+
+
+def get_article(slug):
+    try:
+        article = Article.objects.get(slug=slug)
+        return article
+    except Exception:
+        article_not_found()
