+++ authors/apps/articles/views/article_views.py
@@ -1,0 +1,81 @@
+from django.contrib.auth.models import AnonymousUser
+from rest_framework.generics import (
+    ListCreateAPIView, RetrieveUpdateDestroyAPIView)
+from rest_framework.response import Response
+from rest_framework import status
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from authors.apps.core.pagination import StandardPagination
+from ..permissions import IsOwnerOrReadonly
+from ..renderers import ArticleJsonRenderer
+from ..models import Article
+from ..serializers import ArticleSerializers
+from ..utils import get_article, article_not_found
+
+
+class ListCreateArticle(ListCreateAPIView):
+    queryset = Article.objects.all()
+    renderer_classes = (ArticleJsonRenderer,)
+    serializer_class = ArticleSerializers
+    permission_classes = (IsAuthenticatedOrReadOnly,)
+
+    def get(self, request):
+        paginator = StandardPagination()
+        result_page = paginator.paginate_queryset(self.queryset, request)
+        serializer = ArticleSerializers(result_page, many=True)
+        return paginator.get_paginated_response(serializer.data)
+
+    def post(self, request):
+        article = request.data.get('article', {})
+        serializer = self.serializer_class(
+            data=article, context={'request': request})
+        serializer.is_valid(raise_exception=True)
+        serializer.save(author=request.user)
+
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+
+class RetrieveUpdateDeleteArticle(RetrieveUpdateDestroyAPIView):
+    """
+    This class retrieves, updates, and deletes a single article
+    """
+    queryset = Article.objects.all()
+    serializer_class = ArticleSerializers
+    renderer_classes = (ArticleJsonRenderer,)
+    permission_classes = (IsAuthenticatedOrReadOnly, IsOwnerOrReadonly)
+    lookup_field = 'slug'
+
+    def get(self, request, slug):
+        article = get_article(slug)
+        if not article:
+            article_not_found()
+        if request.user and not isinstance(request.user, AnonymousUser) and article.author != request.user:
+            article.views += 1
+            article.save()
+        return super().get(request, slug)
+
+    def update(self, request, slug):
+        """
+        This method updates an article
+        """
+        try:
+            article = Article.objects.get(slug=slug)
+        except Exception:
+            article_not_found()
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        article_data = request.data.get('article', {})
+        serializer = self.serializer_class(
+            article, data=article_data, partial=True)
+        if serializer.is_valid():
+            self.check_object_permissions(request, article)
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, slug):
+        """
+        Overide the default django error message after deleting an article
+        """
+        if not get_article(slug):
+            article_not_found()
+        super().delete(self, request, slug)
+        return Response({"message": "Article Deleted Successfully"})
