+++ authors/apps/articles/views/tag_view.py
@@ -1,0 +1,17 @@
+from rest_framework.generics import ListAPIView
+from rest_framework.response import Response
+from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
+from ..models import Tag
+from ..serializers import TagSerializer
+
+
+class TagsView(ListAPIView):
+    queryset = Tag.objects.all()
+    serializer_class = TagSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def list(self, request):
+        data = self.get_queryset()
+        serializer = self.serializer_class(data, many=True)
+        return Response({'tags': serializer.data}, status=status.HTTP_200_OK)
